# server/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.1.0"
  title: Lit Hub API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  Bearer:
    description: "For accessing the API a valid JWT token must be passed in all the queries in the 'Authorization' header. A valid JWT token is generated by the API and returned as answer of a call to the route /api/auth/login giving a valid user & password. The following syntax must be used in the 'Authorization' header: Bearer: xxxxxx.yyyyyyy.zzzzzz"
    type: apiKey
    name: Authorization
    in: 'header'
paths:
...
  /books/latest:
    x-swagger-router-controller: Overview
    get:
      description: "Returns the 10 latest books that users have read sorted by their last modified date"
      # controller method name
      operationId: "getBooks"
      responses:
        "200":
          description: "Array of books"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Resource"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
...
  /articles/latest:
    x-swagger-router-controller: Overview
    get:
      description: "Returns the 10 latest articles that users have read sorted by their last modified date"
      # controller method name
      operationId: "getArticles"
      responses:
        "200":
          description: "Array of articles"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Resource"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
...
  /quotes/latest:
    x-swagger-router-controller: Overview
    get:
      description: "Returns the 10 latest quotes that users have read sorted by their last modified date"
      # controller method name
      operationId: "getQuotes"
      responses:
        "200":
          description: "Array of quotes"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Resource"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"

# complex objects have schema definitions
definitions:
  Error:
    type: object
    required:
    - message
    properties:
      message:
        type: string
      errors:
        type: object
...
  Review:
    type: "object"
    properties:
      _id:
        type: "string"
        description: "Resource ID"
      title:
        type: "string"
        description: "Title of the resource"
      author:
        type: "string"
        description: "Author of the resource"
      link:
        type: "string"
        description: "Resource URL (if it applies)"
      rating:
        type: "number"
        description: "User rating for the resource"
      comments:
        type: "string"
        description: "User's thoughts or comments on the resource"
      created:
        type: "string"
        format: "date-time"
        description: "Date when user submitted the review"
      last_modified:
        type: "string"
        format: "date-time"
        description: "Date when the user last modified the review"
  Resource:
    type: "object"
    properties:
      _id:
        type: "string"
        description: "Resource ID"
      title:
        type: "string"
        description: "Title of the resource"
      quote:
        type: "string"
        description: "Quote by a known or unknown author"  
      author:
        type: "string"
        description: "Author of the resource"
      link:
        type: "string"
        description: "Resource URL (if it applies)"
      reviews:
        type: "array"
        items:
          $ref: "#/definitions/Review"
        description: "List of reviews by users for this resource"
      created:
        type: "string"
        format: "date-time"
        description: "Date when the resource was first submitted"
      last_modified:
        type: "string"
        format: "date-time"
        description: "Date when the resource was last modified"
...